Add CALENDAR query support
This commit adds support for the CALENDAR query type, which renders the
results of a dataview query in an obsidian-calendar-ui calendar.

The argument to the CALENDAR query is a date expression which controls
what day the selected note is rendered on.

Currently, each note is rendered as a dot on the calendar. Hovering or
clicking a calendar day will preview or navigate to the first selected
note on that day. Ideally, we would display the note links instead of
dots, but the calendar plugin isn't able to do this (yet).